==> src/1_ti_clustering/0_SetParams.ipynb <==
import json

user = 'bkb12'
table_prefix = 'test'
diseases = {'Parkinsons': '332.0', 'EssentialTremor': '333.1'}
# , 'Alzheimers': '331.0',
#             'SecondaryParkinsonism': '332.1', 'ALS': '335.20'}
per_disease_case_limit = 1000
control_limit = len(diseases) * 1000
enrollment_to_outcome = 365
pre_period = 365
post_period = 730

save_list = {'user': user, 'table_prefix':table_prefix, 'diseases':diseases, 
             'case_limit':per_disease_case_limit, 'control_limit':control_limit,
             'enrollment': enrollment_to_outcome, 'pre_period': 365, 'post_period':730}

with open('../params.json', 'w') as outfile:
    json.dump(save_list, outfile)
