==> src/1_ti_clustering/3_train_model.ipynb <==
import os
os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"   # see issue #152
os.environ["CUDA_VISIBLE_DEVICES"] = "2"

import pyodbc
import pandas as pd
import time
import pickle as pkl
import numpy as np
import matplotlib.pyplot as plt
import json
from pathlib import Path
from keras import backend as K
import datetime
K.clear_session()

%matplotlib inline

with open ('../params.json') as f:
    params = json.load(f)
    
table_prefix = params['table_prefix']
diseases = params['diseases']
case_limit = params['case_limit']
control_limit = params['control_limit']
min_enrollment = params['enrollment']
user = params['user']

chunk = True


# run_id = table_prefix_ + str(case_limit) + '_' + datetime.datetime.now().strftime("%Y-%m-%d_%H:%M:%S")
# print(run_id)

creds_file = "/home/" + user + "/creds.txt" 
creds = lines = [line.rstrip('\n') for line in open(creds_file)]

connection_string = ("Driver={ODBC Driver 17 for SQL Server};" + 
                     "server=" + creds[0] + ";" +
                     "domain=" + creds[1] + ";" +  
                     "database=" + creds[2] + ";" +
                     "uid=" + creds[3]  + ";" +
                     "pwd=" + creds[4] + ";" +
                     "ssl=require;")

cn = pyodbc.connect(connection_string, autocommit=True)
cursor = cn.cursor()

directory = '../../data/diseaes_replaced' + str(table_prefix) + '_' + str(case_limit)
output_dir = '../../outputs/' + str(table_prefix) + _ + str(case_limit) + '_' +\
             'gd'

if not os.path.exists(output_dir):
    os.makedirs(output_dir)

def get_data(path):
    ''' Returns dataframe with columns: 'path', 'word'.'''
    datadir = Path(path)
    files = [(str(f), f.parts[-1]) for f in datadir.glob('*.csv.gz') if f]
    df = pd.DataFrame(files, columns=['path', 'word'])
    
    return df

back_window = 25
train = get_data(directory)
X = train.path
print(X)
X.to_csv(directory + '/files_df.csv')

from keras.utils import to_categorical
import pickle as pkl

def get_seqs(path, vocab_size):
    #print(path)
    seq = pd.read_csv(path)
    #print(seq.shape)
    
    X = []
    decoder_target = []
    y_time = []
    
    for index, row in seq.iterrows():
        if row['fromIndex'] >= 0 and index < seq.shape[0] - 1:
            X.append(to_categorical(seq.iloc[index+1-back_window:index+1]['InputCode'].values, 
                                    num_classes=vocab_size))
            decoder_target.append(to_categorical(seq.iloc[index+2-back_window:index+2]['InputCode'].values, 
                                 num_classes=vocab_size))
            y_time.append(seq.iloc[index+1]['fromIndex'])
    
    
    # encoder input, decoder input, decoder target, y_time
    return (np.asarray(X), np.asarray(X),
            np.asarray(decoder_target), np.asarray(y_time))

    
def batch_generator(X, batch_size=32):
    vocab_dict = pkl.load(open(directory + '/input_token_index.pkl', 'rb'))
    vocab_size = len(vocab_dict)
    while True:
        # choose batch_size random images / labels from the data
        idx = np.random.randint(0, X.shape[0], batch_size)
        x_file = X[idx]
        
        if str(x_file.values[0]) == 'nan':
            continue
        encoder_input, decoder_input, decoder_target, y_time = get_seqs(str(x_file.values[0]), len(vocab_dict))
        
        for i in range(encoder_input.shape[0] / batch_size):
            if (encoder_input.ndim == 3) and (((i+1)*batch_size)-(i*batch_size)) > 1:
                yield ([encoder_input[(i*batch_size):(i+1)*batch_size],
                        decoder_input[(i*batch_size):(i+1)*batch_size]], 
                       [decoder_target[(i*batch_size):(i+1)*batch_size],
                        y_time[(i*batch_size):(i+1)*batch_size,]])
            else:
                pass 
#                 print(encoder_input.shape, decoder_input.shape, decoder_target.shape, y_time.shape)
#                 print(X[idx])
#                 raise Exception('stop')

# print(get_seqs('../../data/test_1000/seq_1200_1300.csv.gz'))
%time [ei_batch, di_batch], [dt_batch, y_time_batch] = batch_generator(X).next()
print('shapes: ', ei_batch.shape, di_batch.shape, dt_batch.shape, y_time_batch.shape)

from keras.models import Model
from keras.utils import multi_gpu_model
from keras.layers import Input, LSTM, Dense, Lambda, GRU, Embedding, GaussianNoise
from keras import backend as K
from keras.callbacks import EarlyStopping, TensorBoard
from keras_exp.multigpu import print_mgpu_modelsummary

def get_model(num_tokens, latent_dim):
    encoder_inputs = Input(shape=(None, num_tokens), name='encoder_input')
    encoder = GRU(latent_dim, return_state=True, name='encoded')
    encoder_outputs, state_h = encoder(encoder_inputs)
                                       
    # Set up the decoder, using `encoder_states` as initial state.
    decoder_inputs = Input(shape=(None, num_tokens), name='decoder_input')
    noisy_inputs = GaussianNoise(0.2)(decoder_inputs)
    decoder_gru = GRU(latent_dim, return_sequences=True, return_state=True)
    decoder_outputs, _ = decoder_gru(noisy_inputs, initial_state=state_h)
    decoder_dense = Dense(num_tokens, activation='softmax', name='ae')
    decoder_outputs = decoder_dense(decoder_outputs)

    time_internal = Dense(100, activation='relu')(encoder_outputs)
    time_dense = Dense(1, activation='relu', name='time')
    time_output = time_dense(time_internal)

    model = Model([encoder_inputs, decoder_inputs], [decoder_outputs, time_output])

    model.summary()
    return model

from sklearn.model_selection import train_test_split
import tensorflow as tf
from keras.callbacks import ModelCheckpoint

print(tf.__version__)
    
config = tf.ConfigProto()
config.gpu_options.allow_growth = True  # dynamically grow the memory used on the GPU
config.log_device_placement = True  # to log device placement (on which device the operation ran)
                                    # (nothing gets printed in Jupyter, only if you run it standalone)
sess = tf.Session(config=config)
K.set_session(sess)  # set this TensorFlow session as the default session for Keras

print(type(X), X.shape[0]*0.8)
Xt = X[int(X.shape[0]*0.8):]
X = X[:int(X.shape[0]*0.8)]

Xt = Xt.reset_index(drop=True)

print(len(X), len(Xt)) 

fixed_length = 25
fixed_vocab = len(pkl.load(open(directory + '/input_token_index.pkl', 'rb')))
# epochs = 100
latent_dim = 10

model = get_model(fixed_vocab, latent_dim)

losses = {'ae':'categorical_crossentropy', 'time':'mse'}
loss_weights = {'ae':1, 'time':2}
earlystopper = EarlyStopping(monitor='loss', patience=100, verbose=1)
tbCallback = TensorBoard(log_dir='../../logs/gd', write_graph=True)

filepath = output_dir + "/weights.{epoch:02d}-{val_loss:.2f}.hdf5"
checkpoint = ModelCheckpoint(filepath, monitor='val_loss', verbose=1, save_best_only=False, mode='min')

model.compile(optimizer='adam', loss=losses, loss_weights=loss_weights)
model.save(output_dir + '/empty_model.model')

batch_size = 1024
train_gen = batch_generator(X, batch_size=batch_size)
valid_gen = batch_generator(Xt, batch_size=batch_size)
print(len(X), len(Xt))

## import time
start_time = time.time()
model.fit_generator(
    generator=train_gen,
    validation_data=valid_gen,
    validation_steps=10,
    epochs=2000,
    steps_per_epoch=100, #X.shape[0], - we do smaller steps per epoch to see training progress
    use_multiprocessing = True,
    workers = 8,
    callbacks=[earlystopper, checkpoint, tbCallback])
print(time.time()-start_time)

model.save(output_dir + '/trained_model_weights.model')
